refactor: SRP에 따라 요청률 제한시 반환 클래스 분리
